feat: implement allow_list support and update example module for forbid_modular_get_outside_module rule
