+++ example/core_module.dart
@@ -1,0 +1,13 @@
+import 'example_forbid_modular_get_outside_module_rule.dart';
+
+class CoreModule extends Module {
+  @override
+  void binds(Injector i) {
+    i.addFactory<ServiceClass>(() {
+      // OK — this file ends in _module.dart
+      final okFoo = Modular.get<Foo>();
+      okFoo.hashCode;
+      return ServiceClass();
+    });
+  }
+}
